Refactor convergence dispatch, pivoting internals, and H.assemble tree/quadstrat resolution
Merge FNormEstimator/iFNormEstimator into a single dual-dispatch type, simplify
CombinedConvCrit/ConvCrit dispatch, and tighten pivoting internals (abstractpivoting,
combinedpivstrat, filldistance, maxvalue, mimicrypivoting). Restore the ACAᵀ
(column-first) compressor to the public API.

H.assemble now resolves its clustering tree backend automatically from which
packages are loaded (H2Trees.TwoNTree, or a k-means clustered tree via the new
ACAParallelKMeansTrees extension when ParallelKMeans.jl is also loaded), and
accepts a `quadstrat` keyword matching BEAST.assemble's convention, auto-deriving
nearquadstrat/farquadstrat via a new `tofarquadstrat` hook (ACABEAST specializes it
to strip singular-quadrature handling for far/well-separated blocks).

Add a progress bar to far-interaction assembly and clean up kernel-matrix
size/dispatch helpers.

+++ src/hmatrix/nearinteractions.jl
@@ -25,7 +25,7 @@ For well-separated clusters (distance > η × cluster diameter), the interaction
 is computed via ACA compression; otherwise, via direct near-field assembly.
 Typical values: `η ≈ 1.0` to `3.0` depending on required accuracy.
 """
-function isnear(η::Real=Float64(1.0))
+function isnear(η::Real=1.0)
     return IsNearFunctor{typeof(η)}(η)
 end
